added Song Lyrics Extractor
A lyrics extractor is used to take out the lyrics from a song. In our project, the user will provide the song name and the artist name and our program will then extract the lyrics of the song from a great resource called api.lyrics.ovh, using requests and JSON libraries.